ass2/requirements.txt
mljar_supervised==1.1.15
optuna==4.3.0
pandas==2.2.3
scikit_learn==1.6.1
